Release/v1.7.5

+++ ajax/email/credential/delete.php
@@ -4,13 +4,15 @@ use \GiveToken\EmailCredential;
 $success = 'false';
 $data = '';
 if (isset($_SESSION['user_id'])) {
-    $id = (int) $_POST['id'];
-    if ($id > 0) {
-        $EmailCredential = new EmailCredential($id);
-        $deleted = $EmailCredential->delete();
-        if ($deleted) {
-            $success = 'true';
-            $data = array('id'=>$id);
+    if (isset($_POST['id'])) {
+        $id = (int) $_POST['id'];
+        if ($id > 0) {
+            $EmailCredential = new EmailCredential($id);
+            $deleted = $EmailCredential->delete();
+            if ($deleted) {
+                $success = 'true';
+                $data = array('id'=>$id);
+            }
         }
     }
 }
